--- src/Reader.java
@@ -1,8 +1,0 @@
-import java.util.Scanner;
-
-public class Reader {
-    public static Scanner scanner = new Scanner(System.in);
-    public static String getLine() {
-        return scanner.nextLine();
-    }
-}
